requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
python-dotenv>=0.19.0
httpx==0.24.0
discord.py>=2.0.0
pydantic==1.10.7
itsdangerous==2.1.2
aiohttp>=3.8.1
pytz>=2021.3
requests>=2.26.0
pymongo==4.11.3
dnspython==2.7.0
numpy==1.26.3
matplotlib==3.8.2
pycountry==23.12.11
geopandas==0.14.3 